feat: support non-cuda devices for text and vision models (#233)
* feat: support non-cuda devices for text models

This commit adds support of non-cuda pytorch backend devices
to text models. Commit extends existing test to run for the
externally specified device (cuda is a default). Commit verified on
Llama3.2-3B-Instruct model for:
* "cuda" device type on NVidia A10 GPU
* "cpu" device type
* "xpu" device type on Intel Data Center Max Series GPU (PVC)

Co-authored-by: anordin95 <alexander.f.nordin@gmail.com>
Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>

* feat: support non-cuda devices for vision models

This commit adds support of non-cuda pytorch backend devices
to vision models. Commit verified on Llama3.2-11B-Vision-Instruct
model for:
* "cuda" device type on NVidia A10 GPU
* "cpu" device type
* "xpu" device type on Intel Data Center Max Series GPU (PVC)

Note that this commit requires a fix on pytorch side for gloo
torch distributed backend to restore TLS on gloo working threads.

Requires: https://github.com/pytorch/pytorch/pull/142184
Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>

* tests: test cpu and on-device inference

This change modifies a test for the reference inference in a way
that on the cpu inference is always tested and on-device is tested
if device is available (currently checking for cuda and xpu in that
order) or if user explicitly specified DEVICE to test via environment
variable.

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>

---------

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>
Co-authored-by: anordin95 <alexander.f.nordin@gmail.com>