Create MITRE
With each model size, please find:
Pretrained weights: These are base weights that can be fine-tuned, domain-adapted with full flexibility.
Instruct weights: These weights are for the models that have been fine-tuned and aligned to follow instructions. They can be used as-is in chat applications or further fine-tuned and aligned for specific use cases.
Protections model: We offer specialized models tailored specifically to developers' needs to detect and filter problematic content or policy-violating inputs and outputs in text and images.